cloudfunc/publish/requirements.txt
functions-framework==3.*
mysql-connector-python
google-cloud-secret-manager